created prompt component
