+++ app/prompts.json
@@ -1,0 +1,34 @@
+{
+    "exercises": [
+        {
+            "sentence": "The dog barked loudly and chased the squirrel up a tree."
+        },
+        {
+            "sentence": "Emily enjoys reading adventure books and solving puzzles."
+        },
+        {
+            "sentence": "The weather forecast predicts rain showers for tomorrow."
+        },
+        {
+            "sentence": "The United States celebrates Independence Day on July 4th."
+        },
+        {
+            "sentence": "Maria wrote a heartfelt letter to her best friend, expressing her gratitude."
+        },
+        {
+            "sentence": "The teacher explained the math problem step by step to help the students understand."
+        },
+        {
+            "sentence": "The library is a quiet place where you can borrow books and study."
+        },
+        {
+            "sentence": "The soccer game ended in a tie, with both teams scoring two goals each."
+        },
+        {
+            "sentence": "The sunflower is a tall plant with bright yellow petals and a brown center."
+        },
+        {
+            "sentence": "The astronauts landed safely on the moon and conducted experiments."
+        }
+    ]
+}
